Moving component directory and relinking imports fixed storybook search path to match
