+++ app/javascript/components/NavBar/NavBar.jsx
