+++ app/javascript/components/NavBar/index.jsx
