+++ app/javascript/components/ShotDetails/ShotDetails.jsx
