+++ app/javascript/components/ShotDetails/ShotRating.jsx
