+++ app/javascript/components/ShotDetails/ShotRotation.jsx
