+++ app/javascript/components/ShotDetails/ShotType.jsx
