+++ app/javascript/components/ShotDetails/index.jsx
