+++ app/javascript/components/StrategyBoard/IceSurface.jsx
