+++ app/javascript/components/StrategyBoard/Rock.jsx
