+++ app/javascript/components/StrategyBoard/StrategyBoard.jsx
