+++ app/javascript/components/StrategyBoard/index.jsx
