+++ app/javascript/packs/index.jsx
@@ -5,7 +5,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 
-import App from './app'
+import App from '../components/Application'
 
 
 document.addEventListener('DOMContentLoaded', () => {
